rename structure packages
