+++ src/main/java/pl/kupniewski/application/authenticate/User.java
@@ -1,4 +1,4 @@
-package pl.kupniewski.application.usecases.authenticate;
+package pl.kupniewski.application.authenticate;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
