+++ src/main/java/pl/kupniewski/application/authenticate/UserRole.java
@@ -1,4 +1,4 @@
-package pl.kupniewski.application.usecases.authenticate;
+package pl.kupniewski.application.authenticate;
 
 public enum UserRole {
     CUTOMER
