+++ src/main/java/pl/kupniewski/application/deliver/OrderReadyToDeliver.java
@@ -1,8 +1,8 @@
-package pl.kupniewski.application.usecases.deliver;
+package pl.kupniewski.application.deliver;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import pl.kupniewski.application.usecases.order.Order;
+import pl.kupniewski.application.order.Order;
 
 @Getter
 @AllArgsConstructor
