+++ src/main/java/pl/kupniewski/presenter/LoginPresenter.java
@@ -3,9 +3,9 @@ package pl.kupniewski.presenter;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
-import pl.kupniewski.application.usecases.authenticate.AuthenticateUser;
-import pl.kupniewski.application.usecases.authenticate.User;
-import pl.kupniewski.application.usecases.authenticate.UserRole;
+import pl.kupniewski.application.authenticate.AuthenticateUser;
+import pl.kupniewski.application.authenticate.User;
+import pl.kupniewski.application.authenticate.UserRole;
 
 @Controller
 @AllArgsConstructor
